Bug[#1278], Task[T472] This implements an idea Robin had.  A small area of L1 scratchpad memory is reserved to hold task-specific information, in particular the stack pointer limit.  This will be used by an upcoming gcc patch to implement stack checking in a way that's compatible with both flat and ELF.  When task switching, this area is swapped out to the thread_info structure.
As a future enhancement, we can add code to libpthread and/or the kernel to
update this information for clone()d tasks; then we'll have stack checking
for threaded applications as well.

Another enhancement will be to let the compiler track the maximum stack usage
by using the second field in the L1 info structure.


git-svn-id: svn://localhost/svn/linux-kernel/trunk@1738 526b6c2d-f592-4532-a319-5dd88ccb003d