[#3283] snipe from -mm tree for OS X users:
[PATCH] remove usage of memmem from scripts/kallsyms.c
The only in-kernel user of "memmem" is scripts/kallsyms.c and it only
uses it to find tokens that are 2 bytes in size. It is trivial to
replace it with a simple function that finds 2-byte tokens.

This should help users from systems that don't have the memmem GNU
extension available.

Signed-off-by: Paulo Marques <pmarques@grupopie.com>
Acked-by: Mike Frysinger <vapier@gentoo.org>


git-svn-id: svn://localhost/svn/linux-kernel/trunk@3320 526b6c2d-f592-4532-a319-5dd88ccb003d