Bug[#3355] Fix a problem with the nommu munmap patch.
libpthread's initialization function assumes that the stack pointer is
pointing at the last page of the stack when it runs.  It uses the value in
SP to set up bounds for the initial thread's stack.

With the recent changes to binfmt_flat, we no longer call ksize on the
object we got when allocating space for bss+stack; we just use the length
we requested with mmap.  This can cause the stack pointer to start out
just above the start of a page, and drop below that page during
pthread_initialize.  The result is that the next time thread_self is called
with the stack pointer back to the page above, it can't find out which
thread is currently running, and the program crashes.

This is fixed by rounding up the length we allocated to the next multiple
of PAGE_SIZE.


git-svn-id: svn://localhost/svn/linux-kernel/trunk@3362 526b6c2d-f592-4532-a319-5dd88ccb003d