[!no_src_qa!] [#2424] Upate kernel to 2.6.22.8
[2.6.22.7]:
We (the -stable team) are announcing the release of the 2.6.22.7 kernel.
It contains a single security bugfix for the x86_64 architecture.
There is potential for local privilege escalation, so all x86_64 users
are certainly encouraged to upgrade.

CVE-2007-4573: x86_64: Zero extend all registers after ptrace in 32bit entry path.

[2.6.22.8]:
We (the -stable team) are announcing the release of the 2.6.22.8 kernel.
It contains a single security bugfix for the the sound subsystem.  There
is potential for local privilege escalation, so all users are certainly
encouraged to upgrade.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@3678 526b6c2d-f592-4532-a319-5dd88ccb003d