[!no_src_qa!] [#2424] This has been added to the mm tree as remove-support-for-un-needed-_extratext-section-checkpatch-fixes.patch No need to send upstream.
This patch is needed on Blackfin to properly decode zero as a null pointer
in arch/blackfin/kernel/traps.c:decode_address()

When passing a zero address to kallsyms_lookup(), the kernel thought it was
a valid kernel address, even if it is not.  This is because is_ksym_addr()
called is_kernel_extratext() and checked against labels that don't exist on
many archs (which default as zero).  Since PPC was the only kernel which
defines _extra_text, (in 2005), and no longer needs it, this patch removes
_extra_text support.

For some history (provided by Jon):
  http://ozlabs.org/pipermail/linuxppc-dev/2005-September/019734.html
  http://ozlabs.org/pipermail/linuxppc-dev/2005-September/019736.html
  http://ozlabs.org/pipermail/linuxppc-dev/2005-September/019751.html

Signed-off-by: Robin Getz <rgetz@blackfin.uclinux.org>
Cc: David Woodhouse <dwmw2@infradead.org>
Cc: Jon Loeliger <jdl@freescale.com>
Cc: Paul Mackerras <paulus@samba.org>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Cc: Sam Ravnborg <sam@ravnborg.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>



git-svn-id: svn://localhost/svn/linux-kernel/trunk@3790 526b6c2d-f592-4532-a319-5dd88ccb003d