[#3719] Rather that duplicate code, connect Hardware Errors (irq5) into trap_c, so user space processes get killed properly.
The rest of irq_panic() can be moved into traps.c (later)

There is still a small corner case that causes problems when a
pheriperal interrupt goes off a single cycle before a user space
hardware error. This causes a kernel panic, rather than the user
space process being killed.

But, this checkin makes things work in 99.9% of the cases, and is a vast
improvement from what is there today (which fails 100% of the time).



git-svn-id: svn://localhost/svn/linux-kernel/trunk@3931 526b6c2d-f592-4532-a319-5dd88ccb003d