[#2424] Update kernel from 2.6.22.15 to 2.6.22.16
We (the -stable team) are announcing the release of the 2.6.22.16
kernel.

It contains a single fix for a problem that could cause a local user to
cause file system corruption on some types of filesystems.

All users of the 2.6.22 series are encouraged to upgrade.

I'll also be replying to this message with a copy of the patch between
2.6.22.15 and 2.6.22.16

The updated 2.6.22.y git tree can be found at:
       git://git.kernel.org/pub/scm/linux/kernel/git/stable/linux-2.6.22.y.git
and can be browsed at the normal kernel.org git web browser:
       http://git.kernel.org/?p=linux/kernel/git/stable/linux-2.6.22.y.git;a=summary

thanks,

greg k-h

--------

 Makefile   |    2 +-
 fs/namei.c |    4 ++--
 2 files changed, 3 insertions(+), 3 deletions(-)

Summary of changes from v2.6.22.15 to v2.6.22.16
================================================

Greg Kroah-Hartman (1):
     Linux 2.6.22.16

Linus Torvalds (1):
     Use access mode instead of open flags to determine needed permissions (CVE-2008-0001)



git-svn-id: svn://localhost/svn/linux-kernel/trunk@4099 526b6c2d-f592-4532-a319-5dd88ccb003d