[#3626] Just let i2c-bfin-twi driver depends on BLACKFIN as Jean's said
Well, there's nothing wrong with letting people build this driver even
if the platform doesn't have any supported device. As long as it
doesn't cause a build failure, it's alright. So you could simply use
"depends on BLACKFIN" (which is technically correct) and just document
which machines have the device (as you already did).

And this is same way as other i2c bus drivers do.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@4485 526b6c2d-f592-4532-a319-5dd88ccb003d