Workaround for DMA mode 1 data corruption anomaly.
Since this anomaly is not announce offically, I just follow the draft
to add this workaround.

Data corruption when using USB DMA mode 1. (Issue manager 17-01-0105)
DMA mode 1 allows large size transfers to generate a single interrupt
at the end of the entire transfer.  The transfer is split up in packets
of length specified in the Maximum Packet Size field for that endpoint.
If the transfer size is not an integer multiple of the Maximum Packet
Size, a short packet will be present at the end of the transfer.

Under certain conditions this packet may be corrupted in the USB FIFO.

Workaround:
Use DMA mode 1 to transfer (n* Maximum Packet Size) and schedule DMA
mode 0 to transfer the short packet.

As an example if your transfer size is 33168 bytes and Maximum Packet
Size equals 512, schedule [33168 - (33168 mod 512)] in DMA mode 1 and
the remainder (33168 mod 512) in DMA mode 0.

Applies to Revisions: 0.0, 0.1 and 0.2 (BF54x)



git-svn-id: svn://localhost/svn/linux-kernel/trunk@5824 526b6c2d-f592-4532-a319-5dd88ccb003d