Fix bug: low byte of w_index is the usb interface number not the whole 2 bytes of w_index
In some usb gadget driver, for example usb audio class device, the high byte of w_index
is the entity id and low byte is the interface number. If we use the 2 bytes of w_index
as the array number, we will get a wrong pointer or NULL pointer.

This patch fixes this issue.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@5872 526b6c2d-f592-4532-a319-5dd88ccb003d