From: Enrik Berkhan <Enrik.Berkhan@ge.com>
The SPORT_HYS bit is not set for BF561 0.5.

IMHO the setting should depend on ANOMALY_05000305 which is about the
availability of the bit, not ANOMALY_05000265 which only describes the
SPORT sensitivity to noise (checked for BF561 only, though).

If that's not true for other BF variants, maybe the definition of
ANOMALY_05000265 for BF561 should be changed to '(1)' instead.

Signed-off-by: Enrik Berkhan <Enrik.Berkhan@ge.com>


git-svn-id: svn://localhost/svn/linux-kernel/trunk@6104 526b6c2d-f592-4532-a319-5dd88ccb003d