Fix bug[#5063] Irregular DMA reg status should be taken care of in auto start mode.
This bug is caused by irregular behavior of DMA register CURR_X_COUNT
and CURR_Y_COUNT when an auto restart uart rx DMA run to last byte in
DMA buffer, trigger the interrupt and stay at this possiton. The status
of current x and y is 0:7 instead of 512:8 or 0:8. The driver doesn't
take care of this case when calculating the position.




git-svn-id: svn://localhost/svn/linux-kernel/trunk@6289 526b6c2d-f592-4532-a319-5dd88ccb003d