netdev feedback from David Miller: Revert to IRQF_SHARED or explain.
Let's do both 

The Blackfin MAC DMA RX interrupt is a dedicated internal systems
interrupt there is now way it can be SHARED with anything else in the
system. In case this interrupt triggers we know its origin. For no
reason we set the IRQF_SHARED flag so that CONFIG_DEBUG_SHIRQ will bug
us here.

Check the source of the interrupt in bfin_mac_interrupt ISR and return
IRQ_NONE if applicable.





git-svn-id: svn://localhost/svn/linux-kernel/trunk@6497 526b6c2d-f592-4532-a319-5dd88ccb003d