[#5146] Develop Linux driver for ADXL345/346 Three-Axis Digital Accelerometers:
Fixup opened logic.
Add FIFO DELAY.

To ensure that the FIFO has
completely popped, there must be at least 5 us between the end
of reading the data registers, signified by the transition to
register 0x38 from 0x37 or the CS pin going high, and the start
of new reads of the FIFO or reading the FIFO_STATUS
register. For SPI operation at 1.5 MHz or lower, the register
addressing portion of the transmission is sufficient delay to
ensure the FIFO has completely popped. It is necessary for SPI
operation greater than 1.5 MHz to de-assert the CS pin to
ensure a total of 5 us, which is at most 3.4 us at 5 MHz
operation.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@6580 526b6c2d-f592-4532-a319-5dd88ccb003d