[#5146] Develop Linux driver for ADXL345/346 Three-Axis Digital Accelerometers
From: Chris Verges <chrisv@cyberswitching.com>

I'd like to propose a slight change to the Kconfig for the adxl34x
driver.  The original Kconfig listed two entries in the
drivers/input/misc menu:

   [ ] ADXL345/346 Three-Axis Digital Accelerometer: I2C Interface
   [ ] ADXL345/346 Three-Axis Digital Accelerometer: SPI Interface

If the I2C option was selected, the SPI option would disappear.  If the
SPI option was selected, the I2C interface would still show.  This is in
addition to both lines being compiled as the same driver (adxl34x) with
no way to tell I2C versus SPI apart.

In an attempt to simplify the interface, the changes made to the Kconfig
attached would put a single line in the menu, with a choice prompt to
select the communication method desired:

   [ ] ADXL345/346 Three-Axis Digital Accelerometer
         ADXL345/346 Communication Method (I2C Interface) --->




git-svn-id: svn://localhost/svn/linux-kernel/trunk@6936 526b6c2d-f592-4532-a319-5dd88ccb003d