Allow hardware errors to be caught during early portions of booting, and leave something in the shadow console that people can use to debug their system with. (to be printed out by the bootloader on next reset).
This enables the hardare error interrupts in head.S, allowing us to
find hardware errors when they happen (well, as much as you can with a
hardware error) and prints out the trace if it is enabled. This will
catch errors (like booting the wrong image on a 533) which previously
resulted in a infinite loop/hang, as well as random hardware errors
before setup_arch().

To disable this debug only feature - turn off EARLY_PRINTK.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@6941 526b6c2d-f592-4532-a319-5dd88ccb003d