Bug [#5412]: 1. Clean up workaround and reference to ANOMALY 05000125, since the parts with this anomaly are not supported.
   cli r1;
   .align 8;
   [p0] = r0;        // write to IMEM_CONTROL or DMEM_CONTROL
   ssync;
   sti r1;

But cli/sti is kept (if orignal code use cli/sti) just to be safe (and to avoid possible implications).

2. CPLB need to be disabled/enabled in several places. So define disable(enable)_icplb() disable(enable)_dcplb() in header asm/cplb.h.

3. Use CSYNC before writing to IMEM_CONTROL/DMEM_CONTROL for correct load store ordering, according to PRM (6-7):
"Note since load store ordering is weak (see "Ordering of Loads and Stores" on page 6-69), disabling of CPLBs should be proceeded by a CSYNC."
Old code use SSYNC, this is not necessary.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@7127 526b6c2d-f592-4532-a319-5dd88ccb003d