merge from mainline: lib/checksum.c: use 32-bit arithmetic consistently
The use of 'unsigned long' variables in the 32-bit part of do_csum()
is confusing at best, and potentially broken for long input on 64-bit
machines.

This changes the code to use 'unsigned int' instead, which makes
the code behave in the same (correct) way on both 32 and 64 bit
machines.


git-svn-id: svn://localhost/svn/linux-kernel/trunk@8017 526b6c2d-f592-4532-a319-5dd88ccb003d