from mainline: blackfin,kgdb,probe_kernel: Cleanup probe_kernel_read/write
From: Jason Wessel <jason.wessel@windriver.com>

Blackfin needs it own arch specific probe_kernel_read() and
probe_kernel_write().

This was moved out of the kgdb code and into the
arch/blackfin/maccess.c, because it is a generic kernel api.

The arch specific kgdb.c for blackfin was cleaned of all functions
which exist in the kgdb core that do the same thing after resolving
the probe_kernel_read() and probe_kernel_write().  This also
eliminated the need for most of the #include's.


git-svn-id: svn://localhost/svn/linux-kernel/trunk@8136 526b6c2d-f592-4532-a319-5dd88ccb003d