[!no_src_qa!]Fix bug[#5824] some modules fail to link symbol strcpy
Because string functions are both inline and exported functions and
folder arch/blackfin/lib is configured as a library path in Makefile,
symbols exported in folder lib is not linked into built-in.o but
inlined only. In order to export string symbols in kernel to modules
properly, they should be exported in other folder.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@8145 526b6c2d-f592-4532-a319-5dd88ccb003d