Bug[#5847] bfin mac hw csum doesn't support ipv6 and ipv4 with head options. Apply patch from Jon Kowal<jon.kowal@dspecialists.de>
blackfin checksum calculation is based on the assumption that a standard
IP packet has a header length of 20 bytes which is true for most IPv4
packets. This assumption is not true for IPv6 packets or IPv4 packets
that make use of header options. We must NOT use the blackfin calculated
checksum for those packets.






git-svn-id: svn://localhost/svn/linux-kernel/trunk@8240 526b6c2d-f592-4532-a319-5dd88ccb003d