add from mainline: tracing: Unify arch_syscall_addr() implementations
Most implementations of arch_syscall_addr() are the same, so create a
default version in common code and move the one piece that differs (the
syscall table) to asm/syscall.h.  New arch ports don't have to waste
time copying & pasting this simple function.

The s390/sparc versions need to be different, so document why.


git-svn-id: svn://localhost/svn/linux-kernel/trunk@8386 526b6c2d-f592-4532-a319-5dd88ccb003d