Add a little optimization to strncpy - which decreases boot time by 20%.
When the kernel is booting, with a initramfs - it builds up the filesystem
from a cpio archive, by calling strncpy_from_user, in fs/namei.c:do_getname()
for every file in the archive (which can be lots) with a length of PATH_MAX
(1024). This causes the dest of the strncpy to be padded with many null bytes.

This optimization mostly causes these many null bytes to be padded with
a call to memset(), which is already optimized for filling memory
quickly.

Boot time measured with 'loglevel=0', so UART speed doesn't get in the
way.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@8686 526b6c2d-f592-4532-a319-5dd88ccb003d