[#5836] run spi flash with dma mode enabled fails to mount.
Some peripherals may generate an error interrupt shortly after the 
Data Interrupt, due to the fact that the peripheral isn't serviced fast
enough.
In most cases this doesn't harm and is expected behavior.
On Blackfin priority driven nested interrupts the Error interrupt 
(with the current priority) overtakes the Data Interrupt. 
This hasn't been a problem on most parts since you simply don't request
the 
ERROR interrupt (or leave it disabled while there is an expected
ERROR/STATUS 
indication) and do the peripheral status checking in the data interrupt.
On BF537/6/4 the ERROR interrupt can't be enabled on peripheral basis, 
once you enable the ERROR interrupt for one peripheral, it automatically 
enables it for all peripherals.

Therefore lower the multiplexed ERROR interrupt priority
and deal with ERROR/STATUS flags in the Data Interrupt.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@8747 526b6c2d-f592-4532-a319-5dd88ccb003d