FLAT: split the stack & data alignments
The stack and data have different alignment requirements, so don't force
them to wear the same shoe.  Increase the data alignment to match that
which the elf2flt linker script has always been using: 0x20 bytes.  Not
only does this bring the kernel loader in line with the toolchain, but
it also fixes a swath of gcc tests which try to force larger alignment
values but ultimately fail.

(this reverts Jie's previous ARCH_FLAT_DATA_ALIGN patch and supports
that requirement a different way)


git-svn-id: svn://localhost/svn/linux-kernel/trunk@8862 526b6c2d-f592-4532-a319-5dd88ccb003d