Task[#6035] reclaim TX skb as early as possible after EMAC tx transfer
SKBs hold onto resources that can't be held indefinitely, such as TCP
socket references and netfilter conntrack state.  So if a packet is left
in TX ring for a long time, there might be a TCP socket that cannot be
closed and freed up.

Current blackfin EMAC driver always reclaim and free used tx skbs in
future transfers. The problem is that future transfer may not come as soon
as possible. This patch start a timer after transfer to reclaim and free
skb. There is nearly no performance drop with this patch.

If Blackfin EMAC TX transfer control is turned on, endless TX
interrupts are triggered no matter if TX DMA is enabled. Since DMA walks
down the ring automatically, TX transfer control can't be turned off in
the middle. The only way is to disable TX interrupt completely.




git-svn-id: svn://localhost/svn/linux-kernel/trunk@8893 526b6c2d-f592-4532-a319-5dd88ccb003d