Bug [#5476] [#5477] force to execute soft reset
it is sure some hardware status are not restored to normal by raise 1,
at least we have been sure uart is not normal and can be restored by
set ier to 0 manually in uboot.
we have no much time to figure out all hardware status. So, force to run
soft reset.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@9009 526b6c2d-f592-4532-a319-5dd88ccb003d