Fix bug[#6189] kgdb: return correct thread id when single step over a breakpoint on current cpu.
Switch to current cpu is allowed in kgdb SMP kernel. If 
1) kernel traps into kgdb at a breakpoint;
2) gdb asks to switch to current cpu thread (-cpu_id -2);
3) gdb asks to continue;

Hc, s and c gdb ops are issued in sequence to skip the breakpoint with
current cpu thread id. But, kgdb always returns the thread id of current
process after the single step is done, which confuses gdb. This patch
return the thread id which Hc is issued with.




git-svn-id: svn://localhost/svn/linux-kernel/trunk@9136 526b6c2d-f592-4532-a319-5dd88ccb003d