bfin_sport_spi: setback default speed after every changed-speed transfer.
As mike pointed out, the SPORT/SPI needs to use the per-transfer speed *only
if it has been specified*, but it needs to do it for that transfer only.
then it must use the default speed set up by the bfin_sport_spi_setup
function for everything else.

This problem was introduced by my last commit #9137, this commit fix it.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@9138 526b6c2d-f592-4532-a319-5dd88ccb003d