[#6210]bfin_latencytest: add a simple driver for preempt latency test
bsong have added preempt to our kernel in commit 9079, add a simple driver
in order to test it.

This driver use TIMER2 to generate interrupt every seconds,the timer isr
handle will record the time and send SIGIO to userapp preempt_test which will
read the recorded time when scheduled and compare with the actual time to
calculate the latency.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@9153 526b6c2d-f592-4532-a319-5dd88ccb003d