vmlinux.lds.h: gather .data..shared_aligned sections in DATA_DATA
With the recent change "net: remove time limit in process_backlog()",
the softnet_data variable changed from "DEFINE_PER_CPU()" to
"DEFINE_PER_CPU_ALIGNED()" which moved it from the .data section to
the .data.shared_align section.  I'm not saying this patch is wrong,
just that is what caused me to notice this larger problem.  No one
else in the kernel is using this aligned macro variant, so I imagine
that's why no one has noticed yet.

Since .data.shared_align isnt declared in any vmlinux files that I can
see, the linker just places it last.  This "just works" for most
people, but when building a ROM kernel on Blackfin systems, it causes
section overlap errors:
bfin-uclinux-ld.real:
    section .init.data [00000000202e06b8 -> 00000000202e48b7] overlaps
    section .data.shared_aligned [00000000202e06b8 -> 00000000202e0723]
I imagine other arches which support the ROM config option and thus do
funky placement would see similar issues ...

On x86, it is stuck in a dedicated section at the end:
 [8] .data             PROGBITS ffffffff810ec000 2ec0000303a8 00 WA 0 0 4096
 [9] .data.shared_alig PROGBITS ffffffff8111c3c0 31c3c00000c8 00 WA 0 0 64

So mark sure we include this section in the DATA_DATA macro so that
it is placed in the right location.


git-svn-id: svn://localhost/svn/linux-kernel/trunk@9241 526b6c2d-f592-4532-a319-5dd88ccb003d