merge from upstream: staging: iio: sca3000 move to new event code scheme and add combined orientations
From: Jonathan Cameron <jic23@cam.ac.uk>

The combined orientations allow handling of typical motion and free fall
detectors to be handled as threshold events just applied to a number of
axes via boolean operations.  So freefall is when x, y and z magnitudes
are all below a threshold for a particular period.  Motion detectors
are typically whether x or y or z are above a particular value.

Signed-off-by: Jonathan Cameron <jic23@cam.ac.uk>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>

git-svn-id: svn://localhost/svn/linux-kernel/trunk@9372 526b6c2d-f592-4532-a319-5dd88ccb003d