[#5426] sir_dev: drop Blackfin anomaly 05000447 handling (rev 7133)
Anomaly 05000447 may exhibit itself at different speeds and not just
115200 baud.  The exact rates where things fail is highly dependent
upon the device the Blackfin is talking to.  So it doesn't make any
sense to restrict 115200 baud rates when lower rates may fail with
some devices and higher rates may work with other devices.

Plus, this change is hacked into common code instead of a Blackfin
driver.  Since this is an issue specific to the bfin_5xx serial
driver when using the IRDA line discipline (ldisc), the change prob
should have lived there.


git-svn-id: svn://localhost/svn/linux-kernel/trunk@9452 526b6c2d-f592-4532-a319-5dd88ccb003d