Bug[#6333] blackfin:kgdb: apply workaround for anomaly 257 to SMP kernel
When run kgdb testing, it looks coreb hangs in sngle step or trap
exception without handling anomaly 05000257 properly on bf561 v0.5.
But, the anomaly list says it apply to bf561 v0.4 and bellow. Apply
its work around to 0.5 temporarily till the behavior and the root
cause are confirmed by the hardware team.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@9489 526b6c2d-f592-4532-a319-5dd88ccb003d