[#6366] drivers:musb: revert commit: 8904 set the NAK limitation according to the device type
Commit 8904 is used for setting NAK limitation for bulk and control transfers
for some bad-behaved peripherals, but it will cause some right-behaved
peripherals such as usb bluetooth can't be detected.

We decided to not support bad-behaved peripherals currently in order to make
right-behaved devices work fine, so revert this patch.



git-svn-id: svn://localhost/svn/linux-kernel/trunk@9515 526b6c2d-f592-4532-a319-5dd88ccb003d