scripts/checkpatch.pl: reset rpt_cleaners warnings
If you run checkpatch against multiple patches, and one of them has a
whitespace issue which can be helped via a script (rpt_cleaners), you
will see the same NOTE over and over for all subsequent patches.  It
makes it seem like those patches also have whitespace problems when in
reality, there's only one or two bad apples.

So reset rpt_cleaners back to 0 after we've issued the note so that it
only shows up near the patch with the actual problems.


git-svn-id: svn://localhost/svn/linux-kernel/trunk@9555 526b6c2d-f592-4532-a319-5dd88ccb003d