[!no_src_qa!][#6458] fs:nfs: Don't use vm_map_ram() in readdir
This patch is from linux kernel tree
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;
h=6650239a4b01077e80d5a4468562756d77afaa59.

Our trunk hasn't include it currently, so import it by hand for some ltp
tests.

===========
vm_map_ram() is not available on NOMMU platforms, and causes trouble
on incoherrent architectures such as ARM when we access the page data
through both the direct and the virtual mapping.

The alternative is to use the direct mapping to access page data
for the case when we are not crossing a page boundary, but to copy
the data into a linear scratch buffer when we are accessing data
that spans page boundaries.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Tested-by: Marc Kleine-Budde <mkl@pengutronix.de>
Cc: stable@kernel.org [2.6.37]



git-svn-id: svn://localhost/svn/linux-kernel/trunk@9582 526b6c2d-f592-4532-a319-5dd88ccb003d