Blackfin: bf537: demux port H mask A and emac rx ints
The BF537 SIC combines the gpio port H mask A interrupts with the
emac rx interrupt, so we need to demux this in software.

It also combines the gpio port H mask B and the emac tx interrupts,
and the watchdog and port F mask B interrupts, but since we don't
support mask B yet, just add the defines for now.


git-svn-id: svn://localhost/svn/linux-kernel/trunk@9854 526b6c2d-f592-4532-a319-5dd88ccb003d