Revert "asm-generic/mutex-dec.h: add SMP support"
This reverts commit 650f1efa82134043f7d39a8db962986f72fdcb95.

After talking to mainline, the SMP barrier logic we had in our Blackfin
code seems to just date back to an older asm-generic/mutex-dec.h from the
initial SMP port.  Common code moved on, and it shouldn't be an issue for
us to do so as well since the atomic funcs that this file builds on have
explicit core_lock grabbing buried in them.  Plus, the system boots for
me without them, so that's a good sign.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>