bug[#6959] kgdb: KGDB SMP IPI notification needs to call generic_exec_single directly.
Current smp_call_function_single generic API is changed to avoid raise IPI and call
function in IPI handler on the same core. But this is necessary to support KGDB to
switch master core in SMP case. So, work around this by removing the static type
of generic_exec_single.

Signed-off-by: Sonic Zhang <sonic.zhang@analog.com>