i2c:i2c-bfin-twi: Read and write the FIFO till it is full.
TWI transfer interrupts may be lost when system is heavily handling other
interrupts, while current transfer handler depends on each accurate interrupt
and miss some data in this case. Because there are 2 2-byte fifos in blackfin
twi controller, the occurrence of the data loss can be reduced by reading and
writing till the FIFO is full.

Signed-off-by: Sonic Zhang <sonic.zhang@analog.com>