pinctrl: pinctrl-adi2: Remove the adi specific reservation map.
Blakcfin new gpio IRQ pins doesn't depend on if pin mode is set to peripheral
or gpio. It is not necessary to check the reservation mode of the requested
pin in IRQ chip operation. Because the peripheral pin mode and gpio pin mode
are checked in pinctrl framework. Remove the adi specific reservation map.

Signed-off-by: Sonic Zhang <sonic.zhang@analog.com>